+++ CHANGELOG.md
@@ -8,6 +8,27 @@ and this project adheres to
 
 <!-- ## [Unreleased] -->
 
+## [Unreleased]
+
+### Changed
+
+- renamed `SendEmail` to `CreateEmailBaseOptions`
+- renamed `email.retrieve` to `email.get`
+- renamed `SendEmailResponse` to `CreateEmailResponse`
+- moved batch related stuff to a new module
+- `email.send` now returns a `CreateEmailResponse` instead of an `EmailId`
+- `batch.send` now returns a `Vec<CreateEmailResponse>` instead of a `Vec<EmailId>`
+- `email.send_batch` is now `batch.send`
+- renamed `ApiKeyData` to `CreateApiKeyOptions`
+- `audiences.create` now returns `CreateAudienceResponse` instead of `AudienceId`
+- `contacts.update` now returns `UpdateContactResponse`
+- renamed `domains.DomainData` to `CreateDomainOptions`
+- `domains.update` now returns `UpdateDomainResponse`
+
+### Deleted
+
+- removed ability to configure user agent via `RESEND_USER_AGENT` (this is no longer configuable)
+
 ## [0.4.0] - 2024-05-01
 
 `@martsokha` basically rewrote the entire repository 0_0